Only update files if changed and remove stale files

+++ src/gulp/stubfs.ts
@@ -14,25 +14,41 @@
 
 import * as fs from 'fs';
 import * as path from 'path';
+import { resolve } from 'path';
+import { promisify } from 'util';
+
+const readdir = promisify(fs.readdir);
+const stat = promisify(fs.stat);
 
 export class StubFS {
+  private ready: Promise<string[]>;
   private storePath: string;
   private newFiles: Map<string, string> = new Map();
 
   public constructor(workspacePath: string) {
     this.storePath = this.createStore(workspacePath);
+    this.ready = this.getFiles(this.storePath);
   }
 
   public newFile(filePath: string, contents: string): void {
     this.newFiles.set(filePath, contents);
   }
 
-  public sync(): void {
+  public async sync(): Promise<void> {
+    const allFiles = new Set(await this.ready);
     this.newFiles.forEach((contents, filePath) => {
       const targetPath = path.join(this.storePath, filePath);
       const directory = path.dirname(targetPath);
       fs.mkdirSync(directory, { recursive: true });
-      fs.writeFileSync(targetPath, contents);
+
+      if (!fs.existsSync(targetPath) || fs.readFileSync(targetPath, 'utf8') !== contents) {
+        fs.writeFileSync(targetPath, contents);
+      }
+      allFiles.delete(targetPath);
+    });
+
+    allFiles.forEach((file) => {
+      fs.unlinkSync(file);
     });
     this.newFiles.clear();
   }
@@ -41,5 +57,16 @@ export class StubFS {
     const storeDirectory = path.join(workspacePath, '.apexlink', 'gulp');
     fs.mkdirSync(storeDirectory, { recursive: true });
     return storeDirectory;
+  }
+
+  private async getFiles(dir: string): Promise<string[]> {
+    const subdirs = await readdir(dir);
+    const files = await Promise.all(
+      subdirs.map(async (subdir) => {
+        const res = resolve(dir, subdir);
+        return (await stat(res)).isDirectory() ? this.getFiles(res) : [res];
+      })
+    );
+    return files.flat();
   }
 }
